# Reduce CPU usage and locking in the connection thread loop