The current design uses a `Queue` and times out every `0.1s` to check if the thread should terminate. This wakes up the thread every `0.1` even if there is no work to do. An `strace` of the python process with `-f` will show this behavior as well as the `futex`. 

To reduce the overhead, the `Queue` has been replaced with a [`SimpleQueue`](https://docs.python.org/3/library/queue.html#queue.SimpleQueue) since the minimum python version was bumped to [3.7+](https://github.com/omnilib/aiosqlite/pull/230) (3.8), and the use case does not require complex functionality. `SimpleQueue` avoids all the threading locks. Additionally the queue consumer thread now terminates when it sees the `_STOP_RUNNING_SENTINEL` so it does not have to wake up every `0.1s` to check if it needs to terminate.

This solution builds on #213 (thanks @vovukman) and takes into consideration the sqlalchemy use case detailed in https://github.com/omnilib/aiosqlite/pull/213#issuecomment-1670856398