@bdraco 2024-01-06
Verified thread looks good on the trace
```
2024-01-06 01:42:31.252 CRITICAL (MainThread) [homeassistant.components.profiler] Thread [Thread-4]: File "/usr/local/lib/python3.12/threading.py", line 1030, in _bootstrap
    self._bootstrap_inner()
  File "/usr/local/lib/python3.12/threading.py", line 1073, in _bootstrap_inner
    self.run()
  File "/usr/local/lib/python3.12/site-packages/aiosqlite/core.py", line 107, in run
    tx_item = self._tx.get()
```