@bdraco 2024-01-06
`strace` the process, all the thread switching every 0.1s and futexs are gone now 👍