@bdraco 2024-01-06
Validated this on my production Home Assistant instance.